@@ -9,44 +9,44 @@
     "start-server": "cd dist/web-portal && http-server -a www.skysmack.test -p 4000 -o",
     "ss:c1": "npm run ss:web:localization && ng serve --host client1.skysmack.test --port 4000 --disable-host-check",
     "ss:c1:open": "npm run ss:web:localization && ng serve --host client1.skysmack.test --port 4000 --open --disable-host-check",
-    "ss:init": "npm i && npm run ss:c1:open",
-    "ss": "run-s ss:lerna ss:web:localization ss:c1",
+    "ss:init": "npm run ss:clean && npm i && npm run ss:lerna && npm run build:lib && npm run ss:c1:open",
+    "ss": "run-s ss:lerna build:lib ss:web:localization ss:c1",
     "ss:lerna": "lerna bootstrap && lerna run build:npm",
     "ss:lerna:watch": "lerna run --stream --no-sort --concurrency 10 build:npm:watch",
     "ss:build": "run-s ss:lerna build:lib ss:web:localization build:c1:prod",
     "ss:prod": "run-s ss:lerna build:lib ss:web:localization  build:c1:dev start-server",
-    "ss:clean": "clean.bat",
+    "ss:clean": "cd ../Assorted/scripts && clean.bat",
     "ss:rebuild": "npm run ss:clean && npm i && npm run ss"
   },
   "private": true,
   "dependencies": {
     "@angular-redux/router": "^9.0.0",
     "@angular-redux/store": "^9.0.0",
-    "@angular/animations": "^7.0.2",
-    "@angular/cdk": "^7.0.3",
-    "@angular/common": "^7.0.2",
-    "@angular/compiler": "^7.0.2",
-    "@angular/core": "^7.0.2",
-    "@angular/forms": "^7.0.2",
-    "@angular/http": "^7.0.2",
-    "@angular/material": "^7.0.3",
-    "@angular/material-moment-adapter": "^7.0.3",
-    "@angular/platform-browser": "^7.0.2",
-    "@angular/platform-browser-dynamic": "^7.0.2",
-    "@angular/pwa": "^0.10.6",
-    "@angular/router": "^7.0.2",
-    "@angular/service-worker": "^7.0.2",
-    "@ngx-translate/core": "^11.0.0",
-    "@ngx-translate/http-loader": "^3.0.1",
-    "@redux-offline/redux-offline": "^2.4.0",
+    "@angular/animations": "^7.2.0",
+    "@angular/cdk": "^7.2.1",
+    "@angular/common": "^7.2.0",
+    "@angular/compiler": "^7.2.0",
+    "@angular/core": "^7.2.0",
+    "@angular/forms": "^7.2.0",
+    "@angular/http": "^7.2.0",
+    "@angular/material": "^7.2.1",
+    "@angular/material-moment-adapter": "^7.2.1",
+    "@angular/platform-browser": "^7.2.0",
+    "@angular/platform-browser-dynamic": "^7.2.0",
+    "@angular/pwa": "^0.12.1",
+    "@angular/router": "^7.2.0",
+    "@angular/service-worker": "^7.2.0",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "@redux-offline/redux-offline": "^2.5.1",
     "angular-calendar": "^0.26.4",
+    "core-js": "^2.6.2",
+    "date-fns": "^1.30.1",
     "gulp": "^4.0.0",
     "gulp-merge-json": "^1.3.1",
-    "gulp-plumber": "^1.2.0",
-    "core-js": "^2.5.4",
-    "date-fns": "^1.29.0",
+    "gulp-plumber": "^1.2.1",
     "hammerjs": "^2.0.8",
-    "moment": "^2.22.2",
+    "moment": "^2.23.0",
     "parse-link-header": "^1.0.1",
     "redux": "^4.0.1",
     "redux-observable": "^1.0.0",
@@ -55,32 +55,32 @@
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.4",
-    "@angular-devkit/build-ng-packagr": "~0.11.4",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "^0.12.1",
+    "@angular-devkit/build-ng-packagr": "~0.12.1",
+    "@angular/cli": "~7.2.1",
+    "@angular/compiler-cli": "^7.2.0",
+    "@angular/language-service": "^7.2.0",
+    "@types/jasmine": "~3.3.5",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~10.12.18",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "lite-server": "^2.4.0",
-    "ng-packagr": "^4.4.5",
+    "ng-packagr": "^4.5.0",
     "normalize-scss": "^7.0.1",
-    "npm-run-all": "^4.1.3",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tsickle": ">=0.29.0",
+    "npm-run-all": "^4.1.5",
+    "protractor": "~5.4.2",
+    "ts-node": "~7.0.1",
+    "tsickle": ">=0.34.0",
     "tslib": "^1.9.3",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.5",
+    "tslint": "~5.12.1",
+    "typescript": "~3.2.2",
     "wait-on": "^3.2.0"
   }
 }
