@@ -2,12 +2,8 @@ import { Lodging, LodgingType } from '@skysmack/packages-lodgings';
 import { LodgingTypesValidation } from './ng-lodging-types-validation';
 import { Injectable } from '@angular/core';
 import { Validators } from '@angular/forms';
-import { FormRule } from '@skysmack/ng-ui';
+import { FormRule, Field, FieldTypes, SelectField, FieldsConfig } from '@skysmack/ng-ui';
 import { LocalObject } from '@skysmack/framework';
-import { Field } from '@skysmack/ng-ui';
-import { FieldTypes } from '@skysmack/ng-ui';
-import { SelectField } from '@skysmack/ng-ui';
-import { FieldsConfig } from '@skysmack/ng-ui';
 
 export interface NgLodgingFormDependencies {
     availableLodgingTypes: LocalObject<LodgingType, number>[];
