@@ -1,11 +1,14 @@
 import { NgModule } from '@angular/core';
 
-import { ReducerRegistry, registerLazyEpics } from '@skysmack/redux';
-import { ProductsEpics, productsReducer, productTypesReducer, ProductTypesEpics } from '@skysmack/packages-products';
-import { NgProductsRequests } from './redux/ng-products-requests';
-import { NgProductTypesRequests } from './redux/ng-product-types-requests';
+import { ReducerRegistry } from '@skysmack/redux';
+import { productsReducer, productTypesReducer } from '@skysmack/packages-products';
 import { NgProductsActions } from './redux/ng-products-actions';
 import { NgProductsStore } from './redux/ng-products-store';
+import { NgProductsEpics } from './redux/ng-products-epics';
+import { NgProductTypesEpics } from './redux/ng-product-types-epics';
+import { registerEpics } from '@skysmack/ng-redux';
+import { NgProductTypesActions } from './redux/ng-product-types-actions';
+import { NgProductTypesStore } from './redux/ng-product-types-store';
 
 @NgModule({
   imports: [],
@@ -13,16 +16,20 @@ import { NgProductsStore } from './redux/ng-products-store';
   providers: [
     [
       { provide: 'NgProductsActions', useClass: NgProductsActions },
-      { provide: 'NgProductsStore', useClass: NgProductsStore }
+      { provide: 'NgProductsStore', useClass: NgProductsStore },
+      { provide: 'NgProductTypesActions', useClass: NgProductTypesActions },
+      { provide: 'NgProductTypesStore', useClass: NgProductTypesStore }
     ]
   ]
 })
 export class NgProductsModule {
-  constructor(productsRequests: NgProductsRequests, productTypesRequests: NgProductTypesRequests) {
+  constructor(
+    productsEpics: NgProductsEpics,
+    productTypesEpics: NgProductTypesEpics
+  ) {
     ReducerRegistry.Instance.register('products', productsReducer);
-    registerLazyEpics(new ProductsEpics(productsRequests).epics);
-
     ReducerRegistry.Instance.register('productTypes', productTypesReducer);
-    registerLazyEpics(new ProductTypesEpics(productTypesRequests).epics);
+    registerEpics(productsEpics);
+    registerEpics(productTypesEpics);
   }
 }
