@@ -1,7 +1,14 @@
 import { ProductTypesIndexComponent } from './product-types-index/product-types-index.component';
 import { ProductTypesCreateComponent } from './product-types-create/product-types-create.component';
 import { ProductTypesEditComponent } from './product-types-edit/product-types-edit.component';
+import { DynamicFieldsIndexComponent, DynamicFieldsCreateComponent, DynamicFieldsEditComponent } from '@skysmack/portal-ui';
 import { Routes } from '@angular/router';
+import { DynamicFieldRouteData } from '@skysmack/framework';
+
+const data = {
+    actionToken: 'NgProductTypesActions',
+    storeToken: 'NgProductTypesStore'
+} as DynamicFieldRouteData;
 
 export const productTypesRoutes: Routes = [
     {
@@ -9,20 +16,17 @@ export const productTypesRoutes: Routes = [
             { path: 'edit/:id', component: ProductTypesEditComponent, pathMatch: 'full' },
             { path: 'create', component: ProductTypesCreateComponent, pathMatch: 'full' },
         ]
+    },
+    {
+        path: 'types/fields', component: DynamicFieldsIndexComponent, data, children: [
+            { path: 'create', component: DynamicFieldsCreateComponent, pathMatch: 'full', data },
+            { path: 'edit/:id', component: DynamicFieldsEditComponent, pathMatch: 'full', data }
+        ]
     }
-    // {
-    //     path: 'types/fields', component: ProductTypesFieldsIndexComponent, children: [
-    //         { path: 'create', component: ProductTypesFieldsCreateComponent, pathMatch: 'full' },
-    //         { path: 'edit/:key', component: ProductTypesFieldsEditComponent, pathMatch: 'full' }
-    //     ]
-    // },
 ];
 
 export const productTypesComponents: any[] = [
     ProductTypesIndexComponent,
     ProductTypesEditComponent,
     ProductTypesCreateComponent,
-    // ProductTypesFieldsIndexComponent,
-    // ProductTypesFieldsCreateComponent,
-    // ProductTypesFieldsEditComponent,
 ];
