@@ -2,7 +2,7 @@ import { Routes } from '@angular/router';
 import { ProductsIndexComponent } from './products-index/products-index.component';
 import { ProductsCreateComponent } from './products-create/products-create.component';
 import { ProductsEditComponent } from './products-edit/products-edit.component';
-import { DynamicFieldsIndexComponent, DynamicFieldsCreateComponent } from '@skysmack/portal-ui';
+import { DynamicFieldsIndexComponent, DynamicFieldsCreateComponent, DynamicFieldsEditComponent } from '@skysmack/portal-ui';
 import { DynamicFieldRouteData } from '@skysmack/framework';
 
 const data = {
@@ -20,7 +20,8 @@ export const productsRoutes: Routes = [
   },
   {
     path: 'fields', component: DynamicFieldsIndexComponent, data, children: [
-      { path: 'create', component: DynamicFieldsCreateComponent, pathMatch: 'full', data }
+      { path: 'create', component: DynamicFieldsCreateComponent, pathMatch: 'full', data },
+      { path: 'edit/:id', component: DynamicFieldsEditComponent, pathMatch: 'full', data }
     ]
   }
 ];
