@@ -3,4 +3,4 @@ export * from './record-form-component';
 export * from './field-form-component';
 export * from './document-record-index-component';
 export * from './document-record-form-component';
-
+export * from './record-index-time-paged-component';
