@@ -26,8 +26,8 @@ export class RecordFormComponent<TAppState, TRecord extends Record<TKey>, TKey, 
     }
 
     ngOnInit() {
-        super.ngOnInit();
         this.editorNavService.showEditorNav();
+        super.ngOnInit();
     }
 
     ngOnDestroy() {
