@@ -3,7 +3,7 @@ import { PackageManifest } from '@skysmack/ng-ui';
 import { PackageLoader } from '@skysmack/ng-packages';
 
 export class MaintenancePackageManifest extends MaintenanceType implements PackageManifest {
-    public icon = 'face';
+    public icon = 'build';
     public menuLocation = 'main';
     public modulePath = './../packages/modules/maintenance_wrapper.module#MaintenanceWrapperModule';
 }
