@@ -1,9 +1,9 @@
-import { NgModule } from "@angular/core";
-import { PackagesModule } from "@skysmack/portal-packages";
+import { NgModule } from '@angular/core';
+import { PackagesModule } from '@skysmack/portal-packages';
 
 @NgModule({
   imports: [
     PackagesModule
   ]
 })
-export class PackagesWrapperModule {}
+export class PackagesWrapperModule { }
