@@ -1,9 +1,9 @@
-import { NgModule } from "@angular/core";
-import { ProductsModule } from "@skysmack/portal-packages";
+import { NgModule } from '@angular/core';
+import { ProductsModule } from '@skysmack/portal-packages';
 
 @NgModule({
   imports: [
     ProductsModule
   ]
 })
-export class ProductsWrapperModule {}
+export class ProductsWrapperModule { }
