@@ -1,6 +1,6 @@
 import { ReduxAction, AppState } from '@skysmack/redux';
 
-export function portalReducer(state: AppState = {}, action: ReduxAction) {
+export function portalReducer(state: AppState = {}, action: ReduxAction): AppState {
 
     switch (action.type) {
         default:
