@@ -1,13 +1,14 @@
 import { NgReduxRouter } from '@angular-redux/router';
 import { NgRedux } from '@angular-redux/store';
 import { createOffline } from '@redux-offline/redux-offline';
-import { applyMiddleware, compose, createStore, DeepPartial, Store, combineReducers, AnyAction } from 'redux';
+import { applyMiddleware, compose, createStore, DeepPartial, Store, combineReducers, AnyAction, Reducer } from 'redux';
 import { createEpicMiddleware, EpicMiddleware } from 'redux-observable';
 import { ReduxOfflineConfiguration } from './redux-offline.configuration';
-import { ReducerRegistry, epic$ } from '@skysmack/redux';
+import { ReducerRegistry } from '@skysmack/redux';
 import { portalReducer } from './portal-reducer';
 import { hydratedReducer } from './hydrated-reducer';
-import { Reducer } from 'redux';
+import { mergeMap } from 'rxjs/operators';
+import { epic$ } from '@skysmack/ng-redux';
 
 export const configureRedux = (ngRedux: NgRedux<any>, ngReduxRouter: NgReduxRouter, reduxOfflineConfiguration: ReduxOfflineConfiguration) => {
     const initialState: DeepPartial<any> = {};
@@ -17,7 +18,7 @@ export const configureRedux = (ngRedux: NgRedux<any>, ngReduxRouter: NgReduxRout
     const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
     // Preserve initial state for not-yet-loaded reducers
-    const combine = (reducers: Reducer<{}, AnyAction>) => {
+    const combine = (reducers: Reducer<{}, AnyAction>): Reducer<{}, AnyAction> => {
         const reducerNames = Object.keys(reducers);
         Object.keys(initialState).forEach(item => {
             if (reducerNames.indexOf(item) === -1) {
@@ -42,13 +43,11 @@ export const configureRedux = (ngRedux: NgRedux<any>, ngReduxRouter: NgReduxRout
     );
 
     // Replace the store's reducer whenever a new reducer is registered.
-    reducerRegistry.setChangeListener(reducers => {
-        store.replaceReducer(combine(reducers));
-    });
+    reducerRegistry.setChangeListener(reducers => store.replaceReducer(combine(reducers)));
 
     ngRedux.provideStore(store);
 
-    epic$.subscribe(rootEpic => epicMiddleware.run(rootEpic));
+    epicMiddleware.run((action$, state$) => epic$.pipe(mergeMap(epic => epic(action$, state$, {}))));
 
     // Enable syncing of Angular router state with our Redux store.
     if (ngReduxRouter) {
