@@ -31,6 +31,10 @@ import { environment } from './../../environments/environment';
         loadChildren: './../packages/modules/packages_wrapper.module#PackagesWrapperModule'
       },
       {
+        path: 'identities',
+        loadChildren: './../packages/modules/identities_wrapper.module#IdentitiesWrapperModule'
+      },
+      {
         path: 'persons',
         loadChildren: './../packages/modules/persons_wrapper.module#PersonsWrapperModule'
       },
@@ -43,8 +47,16 @@ import { environment } from './../../environments/environment';
         loadChildren: './../packages/modules/lodgings_wrapper.module#LodgingsWrapperModule'
       },
       {
+        path: 'lodgingReservations',
+        loadChildren: './../packages/modules/lodging_reservations_wrapper.module#LodgingReservationsWrapperModule'
+      },
+      {
         path: 'oauth2',
         loadChildren: './../packages/modules/oauth2_wrapper.module#OAuth2WrapperModule'
+      },
+      {
+        path: 'maintenance',
+        loadChildren: './../packages/modules/maintenance_wrapper.module#MaintenanceWrapperModule'
       },
       {
         path: '',
@@ -74,7 +86,7 @@ export class StartModule {
   constructor(
     public ngRedux: NgRedux<any>,
     public ngReduxRouter: NgReduxRouter,
-    public reduxOfflineConfiguration: ReduxOfflineConfiguration,
+    public reduxOfflineConfiguration: ReduxOfflineConfiguration
   ) {
     configureRedux(ngRedux, ngReduxRouter, reduxOfflineConfiguration);
   }
