@@ -45,10 +45,11 @@ export abstract class DocumentRecordActionsBase<TStateType, TStore extends Store
 
     public cancelDynamicFieldAction = (field: LocalObject<FieldSchemaViewModel, string>, packagePath: string): void => {
         this.store.dispatch(Object.assign({}, new ReduxAction<CancelDynamicFieldActionPayload<FieldSchemaViewModel>>({
-            type: DocumentRecordActionsBase.CANCEL_DYNAMIC_FIELD_ACTION,
+            type: this.prefix + DocumentRecordActionsBase.CANCEL_DYNAMIC_FIELD_ACTION,
             payload: {
                 field,
-                packagePath
+                packagePath,
+                prefix: this.prefix
             },
             meta: new CancelActionMeta()
         })))
@@ -66,7 +67,7 @@ export abstract class DocumentRecordActionsBase<TStateType, TStore extends Store
     public getSingleField(path: string) {
         this.store.dispatch(Object.assign({}, new ReduxAction<string>({
             type: this.prefix + DocumentRecordActionsBase.GET_SINGLE_FIELD,
-            payload: path
+            payload: this.addAdditionalPaths(path)
         })));
     }
 
@@ -85,7 +86,7 @@ export abstract class DocumentRecordActionsBase<TStateType, TStore extends Store
             meta: {
                 offline: {
                     effect: new Effect<FieldSchemaViewModel[]>(new EffectRequest<FieldSchemaViewModel[]>(
-                        packagePath + '/fields',
+                        this.addAdditionalPaths(packagePath) + '/fields',
                         HttpMethod.POST,
                         fields.map(x => x.object)
                     )),
@@ -116,7 +117,7 @@ export abstract class DocumentRecordActionsBase<TStateType, TStore extends Store
                     // TODO: Add [] FieldSchemaViewModel both places below when fields accepts arrays.
                     effect: new Effect<FieldSchemaViewModel>(new EffectRequest<FieldSchemaViewModel>(
                         // TODO: Use this below when fields accepts array: packagePath + '/fields
-                        packagePath + '/fields/' + fields[0].object.key,
+                        this.addAdditionalPaths(packagePath) + '/fields/' + fields[0].object.key,
                         HttpMethod.PUT,
                         // TODO: Use this below when fields accepts array: fields.map(x => x.object)
                         fields[0].object
@@ -150,7 +151,7 @@ export abstract class DocumentRecordActionsBase<TStateType, TStore extends Store
             meta: {
                 offline: {
                     effect: new Effect<FieldSchemaViewModel[]>(new EffectRequest<FieldSchemaViewModel[]>(
-                        packagePath + '/fields' + paths,
+                        this.addAdditionalPaths(packagePath) + '/fields' + paths,
                         HttpMethod.DELETE,
                         fields.map(x => {
                             x.status = LocalObjectStatus.DELETING
